+++ angular-app/src/app/app.module.ts
@@ -1,12 +1,12 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { AppRoutingModule } from "./app-routing.module";
+import { RoutingAppModule } from "./routing-app.module";
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
 import { AuthComponent } from './auth/auth.component'
 import { AtmosModule } from './atmos-modules/atmos.module';
-import { CoursesModule } from './atmos-apps/courses/courses.module';
+import { CourseCatalogModule } from './atmos-apps/course-catalog/course-catalog.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CookieService } from 'ngx-cookie-service';
 import { MaterialModule } from './material.module';
@@ -23,12 +23,12 @@ import { FlexLayoutModule } from '@angular/flex-layout';
     MaterialModule,
     BrowserModule,
     BrowserAnimationsModule,
-    AppRoutingModule,
-    HttpClientModule,
-    AtmosModule,
-    CoursesModule,
     FormsModule,
     ReactiveFormsModule,
+    RoutingAppModule,
+    HttpClientModule,
+    AtmosModule,
+    CourseCatalogModule,
     HomeLayoutModule,
     SharedModule,
     FlexLayoutModule
