+++ angular-app/src/app/atmos-apps/course-catalog/course-catalog-container/courses-container.component.css
