+++ angular-app/src/app/atmos-apps/course-catalog/course-catalog-container/courses-container.component.html
@@ -50,3 +50,11 @@
     (mouseover)="row.hovered = true" (mouseout)="row.hovered = false"></tr>
 
 </mat-table>
+
+
+
+
+<br><br><br><br><br><br><br>
+<br><br><br><br><br><br><br>
+<br><br><br><br><br><br><br>
+<br><br><br><br><br><br><br>
