+++ angular-app/src/app/atmos-apps/course-catalog/course-details/course-details.component.html
