+++ angular-app/src/app/atmos-apps/course-catalog/course-list/course-list.component.html
