+++ angular-app/src/app/atmos-modules/apps-container/apps-container.component.html
@@ -1,7 +1,7 @@
 <p>Apps Container Works</p>
 
 <button mat-raised-button color="accent" (click)="detectApp()">
-  Curriculum App
+  Course Catalog App
 </button>
 <br><br><br><br><br><br><br><br><br>
 <br><br><br><br><br><br><br><br><br>
