+++ angular-app/src/app/routing-app.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { AuthComponent } from './auth/auth.component';
+
+const appRoutes: Routes = [
+  { path: '', redirectTo: '/auth', pathMatch: 'full' },
+  { path: 'auth', component: AuthComponent },
+  { path: 'home', redirectTo: '/home/apps', pathMatch: 'full' }, //routing-layout
+];
+
+@NgModule({
+  imports: [RouterModule.forRoot(appRoutes, { relativeLinkResolution: 'legacy' })],
+  exports: [RouterModule]
+})
+export class RoutingAppModule { }
